Fix issue with resolving '.deb' packages
When a new cache is opened via 'Cache::debs', any attempt to install said '.deb' packages will fail when running 'Cache::resolve', from what appears to be the system not yet being initialized via 'init_config_system()'. I'm not positive on it, but it appears that if a cache is created without first initing the system, that only packages specified in the argument to 'Cache::debs' are available inside of the cache.

I'm thinking this wasn't caught in the CI since there are tests before the one that tests for debs in 'tests/root.rs', at which point the system would have already been initialized.